Novas classes e funcionalidades em produtos

+++ src/br/com/simplecodes/model/Product.java
@@ -1,16 +1,14 @@
 package br.com.simplecodes.model;
 
-import java.math.BigDecimal;
-
 public class Product {
 	
 	private Long id;
 	private String name;
 	private String description;
-	private BigDecimal price;
+	private Double price;
 	private String type;
 	
-	public Product(Long id, String name, String description, BigDecimal price, String type){
+	public Product(Long id, String name, String description, Double price, String type){
 		this.id = id;
 		this.name = name;
 		this.description = description;
@@ -42,11 +40,11 @@ public class Product {
 		this.description = descrition;
 	}
 
-	public BigDecimal getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
